chore: update import paths after renaming LocationModal #36
